requirements.txt
git+git://github.com/kermitt2/grobid_client_python@master
beautifulsoup4~=4.9.3
pypdf2~=1.26.0
lxml~=4.6.3
pandas~=1.2.4
pdfminer.six~=20201018
requests~=2.25.1
validators~=0.18.2